Restructure Cache to follow new paradigm
