+++ src/Core/Cache/Enum/Type.php
@@ -19,7 +19,7 @@
  *
  */
 
-namespace Friendica\Core\Cache;
+namespace Friendica\Core\Cache\Enum;
 
 /**
  * Enumeration for cache types
