+++ src/Core/Cache/ICache.php
@@ -21,6 +21,8 @@
 
 namespace Friendica\Core\Cache;
 
+use Friendica\Core\Cache\Enum\Duration;
+
 /**
  * Cache Interface
  */
