+++ src/Core/Lock/CacheLock.php
@@ -22,7 +22,7 @@
 namespace Friendica\Core\Lock;
 
 use Friendica\Core\BaseLock;
-use Friendica\Core\Cache\Duration;
+use Friendica\Core\Cache\Enum\Duration;
 use Friendica\Core\Cache\IMemoryCache;
 
 class CacheLock extends BaseLock
