+++ src/Model/APContact.php
@@ -22,7 +22,7 @@
 namespace Friendica\Model;
 
 use Friendica\Content\Text\HTML;
-use Friendica\Core\Cache\Duration;
+use Friendica\Core\Cache\Enum\Duration;
 use Friendica\Core\Logger;
 use Friendica\Core\System;
 use Friendica\Database\DBA;
