+++ src/Model/Photo.php
@@ -21,7 +21,7 @@
 
 namespace Friendica\Model;
 
-use Friendica\Core\Cache\Duration;
+use Friendica\Core\Cache\Enum\Duration;
 use Friendica\Core\Logger;
 use Friendica\Core\System;
 use Friendica\Database\DBA;
