+++ src/Model/Profile.php
@@ -24,7 +24,7 @@ namespace Friendica\Model;
 use Friendica\App;
 use Friendica\Content\Text\BBCode;
 use Friendica\Content\Widget\ContactBlock;
-use Friendica\Core\Cache\Duration;
+use Friendica\Core\Cache\Enum\Duration;
 use Friendica\Core\Hook;
 use Friendica\Core\Logger;
 use Friendica\Core\Protocol;
