+++ tests/src/Core/Lock/ArrayCacheLockTest.php
@@ -21,7 +21,7 @@
 
 namespace Friendica\Test\src\Core\Lock;
 
-use Friendica\Core\Cache\ArrayCache;
+use Friendica\Core\Cache\Type\ArrayCache;
 use Friendica\Core\Lock\CacheLock;
 
 class ArrayCacheLockTest extends LockTest
