+++ tests/src/Core/Lock/MemcacheCacheLockTest.php
@@ -22,7 +22,7 @@
 namespace Friendica\Test\src\Core\Lock;
 
 use Exception;
-use Friendica\Core\Cache\MemcacheCache;
+use Friendica\Core\Cache\Type\MemcacheCache;
 use Friendica\Core\Config\IConfig;
 use Friendica\Core\Lock\CacheLock;
 use Mockery;
